backend/utils/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp_socks==0.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.19.1
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudscraper==1.2.71
cssselect2==0.7.0
curl_cffi==0.7.3
defusedxml==0.7.1
Deprecated==1.2.15
duckduckgo_search==6.3.6
easygoogletranslate==0.0.4
fastapi==0.115.5
Flask==3.1.0
frozenlist==1.5.0
g4f==0.3.6.3
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai @ git+https://github.com/google-gemini/generative-ai-python@2786f33edbabc441b247e549025b3e9e5e849895
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gTTS==2.5.4
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.4
lz4==4.3.3
MarkupSafe==3.0.2
mss==10.0.0
multidict==6.1.0
nest-asyncio==1.6.0
nodriver==0.38
pillow==11.0.0
platformdirs==4.3.6
primp==0.8.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
python-socks==2.5.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tinycss2==1.4.0
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.0
